enable sandbox for renderers that don't run node
