+++ atom/browser/web_contents_preferences.cc
@@ -9,8 +9,6 @@
 #include <vector>
 
 #include "atom/common/native_mate_converters/value_converter.h"
-#include "atom/common/options_switches.h"
-#include "base/command_line.h"
 #include "base/strings/string_number_conversions.h"
 #include "content/public/browser/render_process_host.h"
 #include "content/public/common/content_switches.h"
