+++ atom/renderer/atom_renderer_client.cc
@@ -7,22 +7,21 @@
 #include <string>
 #include <vector>
 
+#include "atom/browser/web_contents_preferences.h"
 #include "atom/common/api/api_messages.h"
 #include "atom/common/api/atom_bindings.h"
 #include "atom/common/api/event_emitter_caller.h"
 #include "atom/common/asar/asar_util.h"
+#include "atom/common/javascript_bindings.h"
 #include "atom/common/native_mate_converters/string16_converter.h"
 #include "atom/common/native_mate_converters/value_converter.h"
 #include "atom/common/node_bindings.h"
 #include "atom/common/node_includes.h"
 #include "atom/common/options_switches.h"
-#include "atom/renderer/api/atom_api_web_frame.h"
+#include "atom/common/native_mate_converters/string16_converter.h"
 #include "atom/renderer/atom_render_view_observer.h"
 #include "atom/renderer/guest_view_container.h"
 #include "atom/renderer/node_array_buffer_bridge.h"
-#include "base/base_paths.h"
-#include "base/command_line.h"
-#include "base/path_service.h"
 #include "chrome/renderer/media/chrome_key_systems.h"
 #include "chrome/renderer/pepper/pepper_helper.h"
 #include "chrome/renderer/printing/print_web_view_helper.h"
@@ -48,149 +47,6 @@
 #if defined(OS_WIN)
 #include <shlobj.h>
 #endif
-
-namespace {
-  // TODO(bridiver) This is copied from atom_api_renderer_ipc.cc
-  // and atom_api_v8_util.cc and should be cleaned up
-  // using
-  // v8::Local<v8::Value> unused = v8::Undefined(isolate);
-  // v8::Local<v8::Object> v8_util = v8::Object::New(isolate);
-  // auto mod = node::get_builtin_module("atom_common_v8_util");
-  // mod->nm_context_register_func(v8_util, unused, context, nullptr);
-  // binding.Set("v8_util", v8_util);
-  // results in an unresolved external symbol error for get_builtin_module
-  // on windows
-  content::RenderView* GetCurrentRenderView() {
-    blink::WebLocalFrame* frame =
-      blink::WebLocalFrame::frameForCurrentContext();
-    if (!frame)
-      return NULL;
-
-    blink::WebView* view = frame->view();
-    if (!view)
-      return NULL;  // can happen during closing.
-
-    return content::RenderView::FromWebView(view);
-  }
-
-  void Send(mate::Arguments* args,
-            const base::string16& channel,
-            const base::ListValue& arguments) {
-    content::RenderView* render_view = GetCurrentRenderView();
-    if (render_view == NULL)
-      return;
-
-    bool success = render_view->Send(new AtomViewHostMsg_Message(
-        render_view->GetRoutingID(), channel, arguments));
-
-    if (!success)
-      args->ThrowError("Unable to send AtomViewHostMsg_Message");
-  }
-
-  base::string16 SendSync(mate::Arguments* args,
-                          const base::string16& channel,
-                          const base::ListValue& arguments) {
-    base::string16 json;
-
-    content::RenderView* render_view = GetCurrentRenderView();
-    if (render_view == NULL)
-      return json;
-
-    IPC::SyncMessage* message = new AtomViewHostMsg_Message_Sync(
-        render_view->GetRoutingID(), channel, arguments, &json);
-    bool success = render_view->Send(message);
-
-    if (!success)
-      args->ThrowError("Unable to send AtomViewHostMsg_Message_Sync");
-
-    return json;
-  }
-
-  v8::Local<v8::Value> GetHiddenValue(v8::Local<v8::Object> object,
-                                    v8::Local<v8::String> key) {
-    return object->GetHiddenValue(key);
-  }
-
-  void SetHiddenValue(v8::Local<v8::Object> object,
-                      v8::Local<v8::String> key,
-                      v8::Local<v8::Value> value) {
-    object->SetHiddenValue(key, value);
-  }
-
-  // TODO(bridiver) This is mostly copied from node_bindings.cc
-  // and should be cleaned up
-  base::FilePath GetResourcesPath() {
-    auto command_line = base::CommandLine::ForCurrentProcess();
-    base::FilePath exec_path(command_line->GetProgram());
-    PathService::Get(base::FILE_EXE, &exec_path);
-
-    base::FilePath resources_path =
-  #if defined(OS_MACOSX)
-        exec_path.DirName().DirName().DirName().DirName().DirName()
-                .Append("Resources");
-  #else
-        exec_path.DirName().Append(FILE_PATH_LITERAL("resources"));
-  #endif
-    return resources_path;
-  }
-
-  // TODO(bridiver) create a separate file for script functions
-  std::string ExceptionToString(const v8::TryCatch& try_catch) {
-    std::string str;
-    v8::HandleScope handle_scope(v8::Isolate::GetCurrent());
-    v8::String::Utf8Value exception(try_catch.Exception());
-    v8::Local<v8::Message> message(try_catch.Message());
-    if (message.IsEmpty()) {
-      str.append(base::StringPrintf("%s\n", *exception));
-    } else {
-      v8::String::Utf8Value filename(message->GetScriptOrigin().ResourceName());
-      int linenum = message->GetLineNumber();
-      int colnum = message->GetStartColumn();
-      str.append(base::StringPrintf(
-          "%s:%i:%i %s\n", *filename, linenum, colnum, *exception));
-      v8::String::Utf8Value sourceline(message->GetSourceLine());
-      str.append(base::StringPrintf("%s\n", *sourceline));
-    }
-    return str;
-  }
-
-  v8::Handle<v8::Value> ExecuteScriptFile(v8::Handle<v8::Context> context,
-                                              base::FilePath script_path) {
-    v8::Isolate* isolate = context->GetIsolate();
-    std::string script_source;
-    asar::ReadFileToString(script_path, &script_source);
-
-    v8::Local<v8::String> source =
-        v8::String::NewFromUtf8(isolate,
-                                script_source.data(),
-                                v8::String::kNormalString,
-                                script_source.size());
-
-    std::string script_name = script_path.AsUTF8Unsafe();
-    v8::Local<v8::String> name =
-        v8::String::NewFromUtf8(isolate,
-                                script_name.data(),
-                                v8::String::kNormalString,
-                                script_name.size());
-
-    v8::TryCatch try_catch;
-    v8::Local<v8::Script> script = v8::Script::Compile(source, name);
-    if (script.IsEmpty()) {
-      LOG(FATAL) << "Failed to parse script file " << script_name;
-      LOG(FATAL) << ExceptionToString(try_catch);
-      exit(3);
-    }
-
-    v8::Local<v8::Value> result = script->Run();
-    if (result.IsEmpty()) {
-      LOG(FATAL) << "Failed to execute script file " << script_name;
-      LOG(FATAL) << ExceptionToString(try_catch);
-      exit(4);
-    }
-
-    return result;
-  }
-}  // namespace
 
 namespace atom {
 
@@ -235,26 +91,22 @@ class AtomRenderFrameObserver : public content::RenderFrameObserver {
 AtomRendererClient::AtomRendererClient()
     : node_bindings_(NodeBindings::Create(false)),
       atom_bindings_(new AtomBindings),
-      run_node_(true) {
-  base::CommandLine* cmd_line = base::CommandLine::ForCurrentProcess();
-  // if node integration is disabled and there is
-  // no preload script then don't start node
-  if ((!cmd_line->HasSwitch(switches::kNodeIntegration) ||
-       cmd_line->GetSwitchValueASCII(switches::kNodeIntegration) == "false") &&
-      !cmd_line->HasSwitch(switches::kPreloadScript) &&
-      !cmd_line->HasSwitch(switches::kPreloadURL)) {
-    run_node_ = false;
-  }
+      javascript_bindings_(new JavascriptBindings) {
 }
 
 AtomRendererClient::~AtomRendererClient() {
+}
+
+// static
+void AtomRendererClient::PreSandboxStartup() {
+  JavascriptBindings::PreSandboxStartup();
 }
 
 void AtomRendererClient::WebKitInitialized() {
   blink::WebCustomElement::addEmbedderCustomElementName("webview");
   blink::WebCustomElement::addEmbedderCustomElementName("browserplugin");
 
-  if (!run_node_)
+  if (!WebContentsPreferences::run_node())
     return;
 
   OverrideNodeArrayBuffer();
@@ -325,7 +177,7 @@ void AtomRendererClient::DidCreateScriptContext(
   // Whether the node binding has been initialized.
   bool first_time = node_bindings_->uv_env() == nullptr;
 
-  if (run_node_) {
+  if (WebContentsPreferences::run_node()) {
     // only load node in the main frame
     if (frame->parent())
       return;
@@ -357,57 +209,13 @@ void AtomRendererClient::DidCreateScriptContext(
     v8::HandleScope handle_scope(isolate);
 
     // Create a process object
-    mate::Dictionary process(isolate, v8::Object::New(isolate));
-
-    // bind native functions
-    mate::Dictionary binding(isolate, v8::Object::New(isolate));
-
-    mate::Dictionary v8_util(isolate, v8::Object::New(isolate));
-    v8_util.SetMethod("setHiddenValue", &SetHiddenValue);
-    v8_util.SetMethod("getHiddenValue", &GetHiddenValue);
-    binding.Set("v8_util", v8_util.GetHandle());
-
-    mate::Dictionary ipc(isolate, v8::Object::New(isolate));
-    ipc.SetMethod("send", &Send);
-    ipc.SetMethod("sendSync", &SendSync);
-    binding.Set("ipc", ipc.GetHandle());
-
-    mate::Dictionary web_frame(isolate, v8::Object::New(isolate));
-    web_frame.Set("webFrame", api::WebFrame::Create(isolate));
-    binding.Set("web_frame", web_frame);
-
-    process.Set("binding", binding);
+    v8::Local<v8::Object> process = v8::Object::New(isolate);
 
     // attach the atom bindings
-    atom_bindings_->BindTo(isolate, process.GetHandle());
+    atom_bindings_->BindTo(isolate, process);
 
-    // store in the global scope
-    v8::Local<v8::String> process_key = mate::StringToV8(isolate, "process");
-    context->Global()->Set(process_key, process.GetHandle());
-
-    // Load everything
-    base::FilePath resources_path = GetResourcesPath();
-    base::FilePath script_path =
-            resources_path.Append(FILE_PATH_LITERAL("atom.asar"))
-                          .Append(FILE_PATH_LITERAL("renderer"))
-                          .Append(FILE_PATH_LITERAL("lib"))
-                          .Append(FILE_PATH_LITERAL("init-without-node.js"));
-    ExecuteScriptFile(context, script_path);
-
-    // Run supplied contentScripts
-    base::CommandLine* cmd_line = base::CommandLine::ForCurrentProcess();
-    if (cmd_line->HasSwitch(switches::kContentScripts)) {
-      std::stringstream ss(cmd_line->
-                              GetSwitchValueASCII(switches::kContentScripts));
-      std::string name;
-      while (std::getline(ss, name, ',')) {
-        base::FilePath file_path;
-        net::FileURLToFilePath(GURL(name), &file_path);
-        ExecuteScriptFile(context, file_path);
-      }
-    }
-    // remove process object from the global scope
-    context->Global()->Delete(process_key);
+    // attach the native function bindings
+    javascript_bindings_->BindTo(isolate, process);
   }
 }
 
@@ -426,7 +234,7 @@ bool AtomRendererClient::ShouldFork(blink::WebLocalFrame* frame,
                                     bool is_initial_navigation,
                                     bool is_server_redirect,
                                     bool* send_referrer) {
-  if (run_node_) {
+  if (WebContentsPreferences::run_node()) {
     *send_referrer = true;
     return http_method == "GET" && !is_server_redirect;
   }
